requirements.txt
flask_sqlalchemy
flask_migrate
flask_login
flask_bootstrap
terra-python
spotipy
